# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: TetraCore Hub API
  description: |
    TetraCore Hub is a distributed task orchestration and management system.

    ## Authentication
    The API supports two authentication methods:
    - **JWT Token**: Use `Authorization: Bearer <token>` header
    - **API Key**: Use `X-API-Key: <key>` header or `Authorization: Bearer tk_<key>`

    ## Rate Limiting
    Default rate limit is 1000 requests per minute. API keys may have custom limits.
  version: 2.0.0
  contact:
    name: TetraCore Support
    email: support@tetra-core.website
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://hub.tetra-core.website
    description: Production
...
paths:
...
  /api/v1/tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      operationId: listTasks
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/status'
        - $ref: '#/components/parameters/priority'
        - $ref: '#/components/parameters/tags'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
    post:
      tags: [Tasks]
      summary: Create task
      operationId: createTask
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
...
  /api/v1/templates:
    get:
      tags: [Templates]
      summary: List templates
      operationId: listTemplates
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Template list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateListResponse'
    post:
      tags: [Templates]
      summary: Create template
      operationId: createTemplate
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: Template created
  /api/v1/templates/{id}/execute:
    post:
      tags: [Templates]
      summary: Execute template
      operationId: executeTemplate
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteTemplateRequest'
      responses:
        '201':
          description: Task created from template
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
    status:
      name: status
      in: query
      schema:
        type: string
        enum: [pending, assigned, processing, completed, failed, cancelled]
    priority:
      name: priority
      in: query
      schema:
        type: string
        enum: [low, normal, high, critical]
    tags:
      name: tags
      in: query
      schema:
        type: string
      description: Comma-separated tags
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
  schemas:
...
    Task:
      type: object
      properties:
        task_id:
          type: string
        task_type:
          type: string
        status:
          type: string
        priority:
          type: string
        payload:
          type: object
        created_at:
          type: string
          format: date-time
    TaskListResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        count:
          type: integer
    CreateTaskRequest:
      type: object
      required: [task_type, payload]
      properties:
        task_type:
          type: string
        payload:
          type: object
        priority:
          type: string
          default: normal
        timeout:
          type: integer
        tags:
          type: array
          items:
            type: string
...
    TemplateListResponse:
      type: object
      properties:
        templates:
          type: array
          items:
            type: object
        count:
          type: integer
    CreateTemplateRequest:
      type: object
      required: [name, task_type, executor_type]
      properties:
        name:
          type: string
        description:
          type: string
        task_type:
          type: string
        executor_type:
          type: string
    ExecuteTemplateRequest:
      type: object
      properties:
        variables:
          type: object
...